# docs/swagger.yaml
definitions:
...
  dto.TokensInput:
    properties:
      grant_type:
        description: 授权类型
        example: client_credentials
        type: string
      scope:
        description: 权限范围
        example: read_write
        type: string
    required:
    - grant_type
    - scope
    type: object
  dto.TokensOutput:
    properties:
      access_token:
        description: access_token
        type: string
      expires_in:
        description: expires_in
        type: integer
      scope:
        description: scope
        type: string
      token_type:
        description: token_type
        type: string
    type: object
  middleware.Response:
    properties:
      data: {}
      errmsg:
        type: string
      errno:
        type: integer
      stack: {}
      trace_id: {}
    type: object
info:
  contact: {}
paths:
...
  /oauth/tokens:
    post:
      consumes:
      - application/json
      description: 获取TOKEN
      operationId: /oauth/tokens
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.TokensInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.TokensOutput'
              type: object
      summary: 获取TOKEN
      tags:
      - OAUTH
...
swagger: "2.0"